refactor: Move PubspecPlus to separate file
- Moved the `PubspecPlus` class and related code from `serverpod_packages_version_check.dart` to a new file `pubspec_plus.dart`.
- Updated imports in `serverpod_packages_version_check.dart` to reference the new `pubspec_plus.dart` file.
- Introduce loadYamlMap function in `yaml_util.dart`. This will correctly parse an empty string as a map, when a map expected (defaults to YamlScalar).
