+++ tools/serverpod_cli/test/integration/serverpod_packages_version_check/serverpod_packages_version_check_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
+import 'package:serverpod_cli/src/serverpod_packages_version_check/pubspec_plus.dart';
 import 'package:serverpod_cli/src/serverpod_packages_version_check/serverpod_packages_version_check.dart';
 import 'package:source_span/source_span.dart';
 import 'package:test/test.dart';
